requirements.txt
amqp==1.4.9
anyjson==0.3.3
backports.ssl-match-hostname==3.5.0.1
billiard==3.3.0.23
celery==3.1.23
cffi==1.8.3
coverage==4.1
cryptography==1.5.2
dj-database-url==0.4.1
Django==1.9.7
django-cache-url==1.0.0
django-filter==0.14.0
django-guardian==1.4.4
django-mailgun==0.9.1
django-nose==1.4.4
django-redis==4.4.3
djangorestframework==3.4.1
djoser==0.5.1
docker-py==1.8.1
ecdsa==0.13
enum34==1.1.6
funcsigs==1.0.2
gunicorn==19.6.0
idna==2.1
ipaddress==1.0.17
kombu==3.0.37
mock==2.0.0
ndg-httpsclient==0.4.2
nose==1.3.7
pbr==1.10.0
psycopg2==2.6.1
pyasn1==0.1.9
pycparser==2.16
pycrypto==2.6.1
pyOpenSSL==16.2.0
pytz==2016.7
redis==2.10.5
requests==2.10.0
six==1.10.0
sshpubkeys==2.2.0
websocket-client==0.37.0
